Struct rustypy::pytypes::pydict::PyDict

pub struct PyDict<K, PyArg> where K: Eq + Hash {
    // some fields omitted
}

Methods

impl<K> PyDict<K, PyArg> where K: Eq + Hash + PyDictKey

fn new() -> PyDict<K, PyArg>

Creates an empty PyDict.

fn insert(&mut self, k: K, v: PyArg) -> Option<PyArg>

Inserts a key-value pair into the map.

If the map did not have this key present, None is returned.

If the map did have this key present, the value is updated, and the old value is returned. The key is not updated, though; this matters for types that can be == without being identical. See the module-level documentation for more.

fn remove(&mut self, k: &K) -> Option<PyArg>

Removes a key from the map, returning the value at the key if the key was previously in the map.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

fn get(&mut self, k: &K) -> Option<&PyArg>

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

fn drain(&mut self) -> Drain<K, PyArg>

Clears the map, returning all key-value pairs as an iterator. Keeps the allocated memory for reuse.

unsafe fn from_ptr(ptr: *mut usize) -> PyDict<K, PyArg>

Get a PyDict from a previously boxed PyDict.

Takes two type parameters: - one corresponding to the key type (check PyArg variants) - and PyArg, corresponding to the value type

And a PyDict as a raw *mut usize pointer.

let dict = unsafe { PyDict::<u64, PyArg>::from_ptr(dict) };

fn as_ptr(self) -> *mut usize

Returns self as raw pointer. Use this method when returning a PyTuple to Python.

fn into_hashmap<V>(self) -> HashMap<K, V> where V: From<PyArg>

Consumes self and returns a HashMap, takes one type parameter and transforms inner content to that type.

fn into_iter<T: From<PyArg>>(self) -> IntoIter<K, T>

Consume self and turn it into an iterator.