Module rustypy::pytypes::pylist

An analog of a Python list which contains elements of a single type, will accept an undefined number of one (and just one) of any other supported type (including other PyLists).

PyList can be constructed from other iterable types as long as the inner type is supported and will own the inner content (a copy will be performed).

PyList::from_iter(vec![1u32; 3]);

You can also use the typical vector interfaces (push, pop, remove, etc.) as long as the type is supported (check PyArg variants).

// required to use push:
use rustypy::pytypes::pylist::PyListPush;

let mut l = PyList::new();
for e in vec![0u32, 1, 3] {
    l.push(e);
}

When extracting in Python with the FFI, elements are moved, not copied (except for PyTuples which require an extra copy) and when free'd all the original elements are dropped.

Unpacking PyList from Python

Is recommended to use the unpack_pylist! macro in order to convert a PyList to a Rust native type. Check the macro documentation for more info.

Structs

PyList
An analog of a Python list which contains an undefined number of elements of a single kind, of any supported type.

Traits

PyListPush

Functions

pylist_extract_pybool

pylist_extract_pydouble

pylist_extract_pyfloat

pylist_extract_pyint

pylist_extract_pylist

pylist_extract_pystring

pylist_extract_pytuple

pylist_free

pylist_len

pylist_new

pylist_push