Struct rustypy::pytypes::pylist::PyList

pub struct PyList {
    // some fields omitted
}

An analog of a Python list which contains an undefined number of elements of a single kind, of any supported type.

Read the module docs for more information.

Methods

impl PyList

fn new() -> PyList

fn remove(&mut self, index: usize) -> PyArg

fn pop(&mut self) -> Option<PyArg>

fn len(&self) -> usize

unsafe fn from_ptr(ptr: *mut PyList) -> PyList

Get a PyList from a previously boxed raw pointer.

fn as_ptr(self) -> *mut PyList

Return a PyList as a raw pointer.