Module rustypy::pytypes::pytuple

An analog of a Python tuple, will accept an undefined number of any other supported types.

You can construct it using the pytuple! macro, ie:

use rustypy::PyArg;
pytuple!(PyArg::I64(10), PyArg::F32(10.5));

You must pass the variety of the argument using the PyArg enum.

When extracting elements in Python with the FFI, elements are copied, not moved, and when free'd all the original elements are dropped.

Unpacking PyTuple from Python

Is recommended to use the unpack_pytuple! macro in order to convert a PyTuple to a Rust native type. Check the macro documentation for more info.

Structs

PyTuple
An analog of a Python tuple, will accept an undefined number of other supported types.

Functions

pytuple_extract_pybool

pytuple_extract_pydouble

pytuple_extract_pyfloat

pytuple_extract_pyint

pytuple_extract_pylist

pytuple_extract_pystring

pytuple_extract_pytuple

pytuple_free

pytuple_len

pytuple_new

pytuple_push