Struct rustypy::pytypes::PyBool

pub struct PyBool {
    pub val: u8,
}

Analog to a Python boolean type.

It supports & and | operators, and comparison to Rust bool types. To return to Python use the as_ptr method and return a raw pointer.

Safety

You can convert a raw pointer to a bool type with from_ptr_into_bool method, or to a &PyBool with from_ method. Those operations are unsafe as they require dereferencing a raw pointer.

Examples

use rustypy::PyBool;
let pybool = PyBool::from(true);
assert_eq(pybool, true);

// prepare to return to Python:
let ptr = pybool.as_ptr();
// convert from raw pointer to a bool
let rust_bool = PyBool::from_ptr_into_bool(ptr);

Fields
val: u8
Methods

impl PyBool

unsafe fn from_ptr(ptr: *mut PyBool) -> &'static PyBool

Dereferences a PyBool raw pointer to an inmutable reference.

unsafe fn from_ptr_into_bool(ptr: *mut PyBool) -> bool

Creates a bool from a raw pointer to a PyBool.

fn to_bool(&self) -> bool

Conversion from PyBool to bool.

fn as_ptr(self) -> *mut PyBool

Returns PyBool as a raw pointer. Use this whenever you want to return a PyBool to Python.