Struct rustypy::pytypes::PyString

pub struct PyString {
    pub ptr: *const c_char,
    pub length: size_t,
}

An analog of a Python String.

To return to Python you must use as_ptr method and return a raw pointer. You can create them using PyString::from trait, from both &str and String.

Safety

When passed from Python you can convert from PyString to an owned string (from_ptr_into_string method) or to a &str slice (to_str method), or to a PyString reference (from_ptr method). Those operations are unsafe as they require dereferencing a raw pointer.

Examples

use rustypy::PyString;
let pystr = PyString::from("Hello world!");

// prepare to return to Python:
let ptr = pystr.as_ptr();
// convert from raw pointer to an owned String
let rust_string = PyString::from_ptr_into_string(ptr);

Fields
ptr: *const c_char length: size_t
Methods

impl PyString

unsafe fn from_ptr(ptr: *mut PyString) -> &'static PyString

Dereferences a PyString raw pointer to an inmutable reference.

fn into_string(&self) -> String

Builds an owned String from a PyString.

fn to_str(&self) -> Cow<str>

Builds a Cow from a PyString.

unsafe fn from_ptr_into_string(ptr: *mut PyString) -> String

Builds an owned String from a raw pointer to a PyString.

fn as_ptr(self) -> *mut PyString

Returns PyString as a raw pointer. Use this whenever you want to return a PyString to Python.