Struct rustypy::pytypes::pybool::PyBool

pub struct PyBool {
    // some fields omitted
}

Analog to a Python boolean type.

Read the module docs for more information.

Methods

impl PyBool

unsafe fn from_ptr(ptr: *mut PyBool) -> PyBool

Get a PyBool from a previously boxed raw pointer.

unsafe fn from_ptr_into_bool(ptr: *mut PyBool) -> bool

Creates a bool from a raw pointer to a PyBool.

fn to_bool(&self) -> bool

Conversion from PyBool to bool.

fn as_ptr(self) -> *mut PyBool

Returns PyBool as a raw pointer. Use this whenever you want to return a PyBool to Python.