Struct rustypy::pytypes::pydict::PyDict

pub struct PyDict<K> where
    K: Eq + Hash + PyDictKey,  { /* fields omitted */ }

Methods

impl<K> PyDict<K> where
K: Eq + Hash + PyDictKey,

fn new() -> PyDict<K>

Creates an empty PyDict.

fn insert<V>(&mut self, k: K, v: V) -> Option<V> where
PyArg: From<V>,
V: From<PyArg>,

Inserts a key-value pair into the map.

If the map did not have this key present, None is returned.

If the map did have this key present, the value is updated, and the old value is returned. The key is not updated, though; this matters for types that can be == without being identical. See the module-level documentation for more.

fn remove<V>(&mut self, k: &K) -> Option<V> where
V: From<PyArg>,

Removes a key from the map, returning the value at the key if the key was previously in the map.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

fn get<'a, V>(&'a mut self, k: &K) -> Option<V> where
V: From<&'a PyArg>,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map's key type, but Hash and Eq on the borrowed form must match those for the key type.

unsafe fn from_ptr(ptr: *mut size_t) -> PyDict<K>

Get a PyDict from a previously boxed PyDict.

Takes two type parameters: - one corresponding to the key type (check PyArg variants) - and PyArg, corresponding to the value type

And a PyDict as a raw *mut size_t pointer.

let dict = unsafe { PyDict::<u64>::from_ptr(dict) };

fn as_ptr(self) -> *mut size_t

Returns self as raw pointer. Use this method when returning a PyTuple to Python.

fn into_hashmap<V>(self) -> HashMap<K, V> where
V: From<PyArg>,

Consumes self and returns a HashMap, takes one type parameter and transforms inner content to that type.

fn into_iter<V: From<PyArg>>(self) -> IntoIter<K, V>

Consume self and turn it into an iterator.