Struct rustypy::pytypes::pylist::PyList

pub struct PyList {
    // some fields omitted
}

An analog of a Python list which contains an undefined number of elements of a single kind, of any supported type.

Read the module docs for more information.

Methods

impl PyList

fn new() -> PyList

Constructs a new, empty Vec<T>.

The vector will not allocate until elements are pushed onto it.

fn remove(&mut self, index: usize) -> PyArg

Removes and returns the element at position index within the vector, shifting all elements after it to the left.

fn pop(&mut self) -> Option<PyArg>

Removes the last element from a vector and returns it, or None if it is empty.

fn len(&self) -> usize

Returns the number of elements in the PyList.

fn push<T>(&mut self, a: T) where PyArg: From<T>

Appends an element to the back of a collection.

Panics

Panics if the number of elements in the vector overflows a usize.

unsafe fn from_ptr(ptr: *mut PyList) -> PyList

Get a PyList from a previously boxed raw pointer.

fn as_ptr(self) -> *mut PyList

Return a PyList as a raw pointer.

fn into_iter<T: From<PyArg>>(self) -> IntoIter<T>

Consume self and turn it into an iterator.